cs4459/cs4459_assignment3/inventory_service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
psycopg2-binary==2.9.9